rename runQ to evalQ to match up with state monad naming.
this is to allow runQ to be used later on in a way that
also matches up with the state monad: evalQ returns the
final answer, runQ returns the final answer and the
database.

There could also then be execQ that only returns the
db, but that is not needed at the moment.

Both execQ and evalQ are projections of runQ