canonicalise RRsets.
There are two places where this doesn't work quite as I expected:

rdlen is different sometimes (perhaps due to compression?) so
the naive "deriving Ord" canonicalisation doesn't work.

Sometimes a result is returned for a query with a . on the end,
sometimes not. I should be a bit more careful about what I am
asking for.